requirements.txt
alembic==0.9.9
Babel==2.6.0
backcall==0.1.0
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.3.0
dominate==2.3.1
elasticsearch==6.3.0
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.0
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.7
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
parso==0.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.8
traitlets==4.3.2
urllib3==1.23
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
